+++ integrations/ai-sdk-provider/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Renamed internal `fmapi-language-model` module to `chat-completions-language-model` for clarity
+- Renamed internal `responses-agent-language-model` module to `responses-language-model` for clarity
 
 ## [0.2.2] - 2026-01-26